requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
Flask==2.0.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dotenv==0.19.0
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
editdistance==0.6.0
lmdb==1.2.1
matplotlib==3.1.1
numpy>=1.18.1
opencv-python==4.5.4.58
path==15.0.0
tensorflow-cpu==2.7.0
pillow== 6.2.0